requirements.txt
# Telegram Bot
aiogram==3.15.0

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Google APIs
google-generativeai==0.8.4
google-auth-oauthlib==1.2.1
google-api-python-client==2.159.0

# Auth
PyJWT==2.10.1

# Settings
pydantic-settings==2.7.1

# Utils
python-dotenv==1.0.1
